discord-soldier-bot/requirements.txt
discord.py
gspread
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2